+++ drift_hrana/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Properly handle connections closing.
+
 ## 1.0.0
 
 - Initial version.