+++ hrana/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Add `Database.closed`, a future that completes when the underlying connection
+  has closed.
+
 ## 0.1.0
 
 Initial version.